Issue#19: SignUp

+++ backend/src/main/java/com/example/backend/service/PetService.java
@@ -1,13 +1,13 @@
 package com.example.backend.service;
 
 import com.example.backend.exception.PetNotFoundException;
+import com.example.backend.exception.UnauthorizedException;
 import com.example.backend.model.Pet;
 import com.example.backend.model.PetDTO;
 import com.example.backend.repository.PetRepo;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,33 +16,44 @@ import java.util.Optional;
 public class PetService {
     private final PetRepo petRepo;
     private final IdService idService;
+    private final MongoUserDetailsService mongoUserDetailsService;
     public List<Pet> getAllPets(){ return petRepo.findAll();}
 
-    public Pet addPet(PetDTO newPet){
+    public Pet addPet(PetDTO newPet, Principal principal){
         if(newPet.name()==null || newPet.name().equals("")){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Pet name is missing or empty");
+            throw new IllegalArgumentException("Pet name is missing or empty");
         }
         String id = idService.generateId();
-        Pet newPetWithId = new Pet(id,newPet.name(),newPet.nameOfBreed(), newPet.photo(),newPet.supplies());
+        Pet newPetWithId = new Pet(id,newPet.name(),newPet.nameOfBreed(), newPet.photo(),newPet.supplies(),mongoUserDetailsService.getMe(principal).id());
         return petRepo.save(newPetWithId);
 
     }
     public Pet getPetById(String id){
         return petRepo.findById(id).orElseThrow(PetNotFoundException::new);
     }
-    public Pet updatePet(String id, PetDTO pet) {
+    public Pet updatePet(String id, PetDTO pet, Principal principal) {
+        String userId= mongoUserDetailsService.getMe(principal).id();
         Optional<Pet> optionalPet =petRepo.findById(id);
         if(optionalPet.isEmpty()){
             throw new PetNotFoundException("Pet with Id: " +id + " doesn't exist");
         }
-        Pet updatedPet=new Pet(id, pet.name(),pet.nameOfBreed(), pet.photo(), pet.supplies());
+        if(!optionalPet.get().userId().equals(userId)){
+            throw new UnauthorizedException("You are only allowed to update your own pets");
+
+        }
+        Pet updatedPet=new Pet(id, pet.name(),pet.nameOfBreed(), pet.photo(), pet.supplies(), userId);
         return petRepo.save(updatedPet);
     }
 
-    public Pet deletePet(String id) {
+    public Pet deletePet(String id, Principal principal) {
         Optional<Pet> optionalPet =petRepo.findById(id);
+        String userId= mongoUserDetailsService.getMe(principal).id();
         if (optionalPet.isEmpty()){
             throw new PetNotFoundException("Pet with id"+ id+ "doesn't exist");
+        }
+        if(!optionalPet.get().userId().equals(userId)){
+            throw new UnauthorizedException("You are only allowed to delete your own pets");
+
         }
         petRepo.deleteById(id);
         return optionalPet.get();
